app/requirements.txt
numpy
psycopg2-binary
dlib==19.22.0
face-recognition==1.3.0